Split UI tests into e2e vs rest_api integration suites
